requirements_dev.txt
codecov==2.0.15
freezegun==0.3.11
flake8==3.7.7
pydocstyle==3.0.0
nose==1.3.7
coverage==4.5.3
mock==3.0.4
pact-python==0.18.0

-r requirements.txt